backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
anthropic==0.18.1
pinecone-client==3.0.3
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
numpy==1.26.4
tiktoken==0.6.0
